Fixed compatibility with 2017.x

+++ PostProcessing/Runtime/PostProcessLayer.cs
@@ -990,7 +990,7 @@ namespace UnityEngine.Rendering.PostProcessing
 
             if (context.stereoActive && context.stereoRenderingMode == PostProcessRenderContext.StereoRenderingMode.SinglePassInstanced)
             {
-                uberSheet.properties.SetInt(ShaderIDs.DepthSlice, eye);
+                uberSheet.properties.SetFloat(ShaderIDs.DepthSlice, eye);
                 cmd.BlitFullscreenTriangleToTexArray(context.source, context.destination, uberSheet, 0, false, eye);
             }
             else if (isFinalPass)
@@ -1025,7 +1025,7 @@ namespace UnityEngine.Rendering.PostProcessing
 
                 if (context.stereoActive && context.stereoRenderingMode == PostProcessRenderContext.StereoRenderingMode.SinglePassInstanced)
                 {
-                    sheet.properties.SetInt(ShaderIDs.DepthSlice, eye);
+                    sheet.properties.SetFloat(ShaderIDs.DepthSlice, eye);
                     cmd.BlitFullscreenTriangleToTexArray(context.source, context.destination, sheet, 0, false, eye);
                 }
                 else
@@ -1068,7 +1068,7 @@ namespace UnityEngine.Rendering.PostProcessing
                 ApplyFlip(context, uberSheet.properties);
                 if (context.stereoActive && context.stereoRenderingMode == PostProcessRenderContext.StereoRenderingMode.SinglePassInstanced)
                 {
-                    uberSheet.properties.SetInt(ShaderIDs.DepthSlice, eye);
+                    uberSheet.properties.SetFloat(ShaderIDs.DepthSlice, eye);
                     cmd.BlitFullscreenTriangleToTexArray(context.source, context.destination, uberSheet, 0, false, eye);
                 }
                 else
